Move files in testrunner to test_helpers to centralize test fixtures
