+++ test_helpers/helpers.go
@@ -1,4 +1,4 @@
-package testrunner
+package test_helpers
 
 import (
 	"code.cloudfoundry.org/routing-api/config"
